+++ R/paginate.R
@@ -48,12 +48,12 @@
 #' @examples
 #' page_size <- 40
 #'
-#' request(example_url()) %>%
-#'   req_url_path("/iris") %>%
-#'   req_url_query(limit = page_size) %>%
+#' request(example_url()) |>
+#'   req_url_path("/iris") |>
+#'   req_url_query(limit = page_size) |>
 #'   req_paginate_page_index(
 #'     page_index = function(req, page) {
-#'       req %>% req_url_query(page_index = page)
+#'       req |> req_url_query(page_index = page)
 #'     },
 #'     parse_resp = function(resp) {
 #'       parsed <- resp_body_json(resp)
@@ -131,12 +131,12 @@ req_paginate <- function(req,
 #' @examples
 #' page_size <- 40
 #'
-#' req_flowers <- request(example_url()) %>%
-#'   req_url_path("/iris") %>%
-#'   req_url_query(limit = page_size) %>%
+#' req_flowers <- request(example_url()) |>
+#'   req_url_path("/iris") |>
+#'   req_url_query(limit = page_size) |>
 #'   req_paginate_page_index(
 #'     page_index = function(req, page) {
-#'       req %>% req_url_query(page_index = page)
+#'       req |> req_url_query(page_index = page)
 #'     },
 #'     parse_resp = function(resp) {
 #'       parsed <- resp_body_json(resp)
@@ -225,17 +225,15 @@ req_perform_iteratively <- function(req,
 #' @return Generates the next request in an iterative request,
 #'   or `NULL` if there are no more pages to return.
 #' @examples
-#' req_flowers <- request(example_url()) %>%
-#'   req_url_path("/iris") %>%
-#'   req_url_query(limit = 40) %>%
+#' req_flowers <- request(example_url()) |>
+#'   req_url_path("/iris") |>
+#'   req_url_query(limit = 40) |>
 #'   req_paginate_page_index(
-#'     page_index = function(req, page) {
-#'       req %>% req_url_query(page_index = page)
-#'     }
+#'     page_index = \(req, page) req |> req_url_query(page_index = page)
 #'   )
 #' req_flowers$url
 #'
-#' resp <- req_flowers %>% req_perform()
+#' resp <- req_flowers |> req_perform()
 #' next_req <- iterate_next_request(req_flowers, resp)
 #' next_req$url
 iterate_next_request <- function(req, parsed) {