+++ R/req-perform-stream.R
@@ -18,8 +18,8 @@
 #'   cat("Got ", length(x), " bytes\n", sep = "")
 #'   TRUE
 #' }
-#' resp <- request(example_url()) %>%
-#'   req_url_path("/stream-bytes/100000") %>%
+#' resp <- request(example_url()) |>
+#'   req_url_path("/stream-bytes/100000") |>
 #'   req_perform_stream(show_bytes, buffer_kb = 32)
 req_perform_stream <- function(req, callback, timeout_sec = Inf, buffer_kb = 64) {
   check_request(req)