+++ R/req-url.R
@@ -9,10 +9,9 @@
 #' @inheritParams req_perform
 #' @param url New URL; completely replaces existing.
 #' @param ... For `req_url_query()`: <[`dynamic-dots`][rlang::dyn-dots]>
-#'   Name-value pairs that provide query parameters. Each value must be either
-#'   a length-1 atomic vector (which is automatically escaped) or `NULL` (which
-#'   is silently dropped). If you want to opt out of escaping, wrap strings in
-#'   `I()`.
+#'   Name-value pairs that define query parameters. Each value must be either
+#'   an atomic vector or `NULL` (which removes the corresponding parameters).
+#'   If you want to opt out of escaping, wrap strings in `I()`.
 #'
 #'   For `req_url_path()` and `req_url_path_append()`: A sequence of path
 #'   components that will be combined with `/`.
@@ -22,19 +21,23 @@
 #' req <- request("http://example.com")
 #'
 #' # Change url components
-#' req %>%
-#'   req_url_path_append("a") %>%
-#'   req_url_path_append("b") %>%
-#'   req_url_path_append("search.html") %>%
+#' req |>
+#'   req_url_path_append("a") |>
+#'   req_url_path_append("b") |>
+#'   req_url_path_append("search.html") |>
 #'   req_url_query(q = "the cool ice")
 #'
 #' # Change complete url
-#' req %>%
+#' req |>
 #'   req_url("http://google.com")
+#'
+#' # Use .multi to control what happens with vector parameters:
+#' req |> req_url_query(id = 100:105, .multi = "comma")
+#' req |> req_url_query(id = 100:105, .multi = "explode")
 #'
 #' # If you have query parameters in a list, use !!!
 #' params <- list(a = "1", b = "2")
-#' req %>%
+#' req |>
 #'   req_url_query(!!!params, c = "3")
 req_url <- function(req, url) {
   check_request(req)
@@ -46,13 +49,78 @@ req_url <- function(req, url) {
 
 #' @export
 #' @rdname req_url
-req_url_query <- function(.req, ...) {
+#' @param .multi Controls what happens when an element of `...` is a vector
+#'   containing multiple values:
+#'
+#'   * `"error"`, the default, throws an error.
+#'   * `"comma"`, separates values with a `,`, e.g. `?x=1,2`.
+#'   * `"pipe"`, separates values with a `|`, e.g. `?x=1|2`.
+#'   * `"explode"`, turns each element into its own parameter, e.g. `?x=1&x=2`.
+#'
+#'   If none of these functions work, you can alternatively supply a function
+#'   that takes a character vector and returns a string.
+req_url_query <- function(.req,
+                          ...,
+                          .multi = c("error", "comma", "pipe", "explode")) {
   check_request(.req)
+  if (is.function(.multi)) {
+    multi <- .multi
+  } else {
+    multi <- arg_match(.multi)
+  }
+
+  dots <- list2(...)
+
+  type_ok <- map_lgl(dots, function(x) is_atomic(x) || is.null(x))
+  if (any(!type_ok)) {
+    cli::cli_abort(
+      "All elements of {.code ...} must be either an atomic vector or NULL."
+    )
+  }
+
+  n <- lengths(dots)
+  if (any(n > 1)) {
+    if (is.function(multi)) {
+      dots[n > 1] <- lapply(dots[n > 1], format_query_param)
+      dots[n > 1] <- lapply(dots[n > 1], multi)
+      dots[n > 1] <- lapply(dots[n > 1], I)
+    } else if (multi == "comma") {
+      dots[n > 1] <- lapply(dots[n > 1], format_query_param)
+      dots[n > 1] <- lapply(dots[n > 1], paste0, collapse = ",")
+      dots[n > 1] <- lapply(dots[n > 1], I)
+    } else if (multi == "pipe") {
+      dots[n > 1] <- lapply(dots[n > 1], format_query_param)
+      dots[n > 1] <- lapply(dots[n > 1], paste0, collapse = "|")
+      dots[n > 1] <- lapply(dots[n > 1], I)
+    } else if (multi == "explode") {
+      dots <- explode(dots)
+    } else if (multi == "error") {
+      cli::cli_abort(c(
+        "All vector elements of {.code ...} must be length 1.",
+        i = "Use {.arg .multi} to choose a strategy for handling vectors."
+      ))
+    }
+  }
+  # Force query generation to bubble up errors
+  query_build(dots)
 
   url <- url_parse(.req$url)
-  url$query <- modify_list(url$query, ...)
-
+  url$query <- modify_list(url$query, !!!dots)
   req_url(.req, url_build(url))
+}
+
+explode <- function(x) {
+  expanded <- map(x, function(x) {
+    if (is.null(x)) {
+      list(NULL)
+    } else {
+      map(seq_along(x), function(i) x[i])
+    }
+  })
+  stats::setNames(
+    unlist(expanded, recursive = FALSE, use.names = FALSE),
+    rep(names(x), lengths(expanded))
+  )
 }
 
 #' @export